+++ src/inventory/Inventory.ts
@@ -8,6 +8,23 @@ import { IK2Template } from "../types/IK2Template";
 
 import { inventoryKind } from "./kinds";
 
+export async function getInventory(inventoryPath: string): Promise<Inventory> {
+  if (!fs.existsSync(inventoryPath)) {
+    throw new Error(
+      "current folder doesn't contains k2.inventor.yaml file at " +
+        inventoryPath
+    );
+  }
+
+  const inventoryFolder = path.dirname(inventoryPath);
+  const inventoryFilename = path.basename(inventoryPath);
+  const inventory = new Inventory(inventoryFilename, inventoryFolder);
+
+  await inventory.load();
+
+  return inventory;
+}
+
 export class Inventory {
   constructor(inventoryFilename: string, inventoryFolder: string) {
     this.inventory_path = path.resolve(inventoryFolder, inventoryFilename);
